pyproject.toml
[project]
name = "knowledge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-allauth>=65.11.1",
    "django-debug-toolbar>=6.0.0",
    "django-environ>=0.12.0",
    "django-import-export>=4.3.9",
    "django-mptt>=0.18.0",
    "django-widget-tweaks>=1.5.0",
    "django>=5.2.6",
    "djlint>=1.36.4",
]
